requirements.txt
torch==2.0
pyyaml
easydict
einops
numpy<2.0
nibabel
pybind11
tqdm
tensorboard
h5py
scikit-learn
open3d==0.11.2